Re-factored the ExperimentSetup by:
- Encapsulating most of its parameters
- Using unordered_map instead of using several `arma::Row`'s
- Replaced most `arma::Row`'s with arma::Mat as they are very similar anyway
- Implemented a few setter function for each variables, they need to be test yet

+++ include/Experiment.h
@@ -30,7 +30,7 @@ namespace sam {
     public:
         ExperimentSetup setup;
 
-        arma::Row<int> nobs;
+        arma::Row<double> nobs;
         arma::Row<double> means;
         arma::Row<double> vars;
         arma::Row<double> ses;
